added del ui and date component
