+++ next-js/styles/Home.module.css
@@ -5,20 +5,14 @@
 .main {
   min-height: 100vh;
   padding: 4rem 0;
-  /* display: flex;
-  flex-direction: column; */
-  /* flex: 1;
-  justify-content: center; */
-  /* align-items: center;  */
 }
 
 .inputLayout {
   display: flex;
   align-items: center;
   gap: 10px;
-  max-width: 500px;
+  justify-content: center;
 }
-
 
 .footer {
   display: flex;
