+++ next-js/styles/input/InputBox.module.css
@@ -1,6 +1,7 @@
 .box {
     width: 100%;
-    padding: 12px 20px;
+    padding: 12px 12px;
     margin: 8px 0;
     box-sizing: border-box;
+    max-width: 400px;
   }
