+++ next-js/styles/input/InputDeleteBtn.module.css
@@ -1,0 +1,28 @@
+.btn {
+    width: 40px;
+    height: 32px;
+    color: white;
+    background-color: transparent;
+    border: none;
+    position: relative;
+    border-right: 1px solid white;
+    border-bottom: 1px solid white;
+    cursor: pointer;
+    caret-color: transparent;
+
+}
+
+.btn:hover {
+    background-color: white;
+    color: black;
+}
+
+
+.btn:active {
+    transform: scale(0.48);
+    background-color: white;
+    color: black;
+}
+
+.btnLabel {
+}
